REFACTOR: changed to eigen float to double

+++ src/reconstruction/ts_plane_tag.hh
@@ -20,16 +20,16 @@ namespace tslam
          * 
          * @return The coordinates per each point
          */
-        void setCorners(std::vector<Eigen::Vector3f> corners);
-        void setCorners(Eigen::Vector3f A, Eigen::Vector3f B, Eigen::Vector3f C, Eigen::Vector3f D);
+        void setCorners(std::vector<Eigen::Vector3d> corners);
+        void setCorners(Eigen::Vector3d A, Eigen::Vector3d B, Eigen::Vector3d C, Eigen::Vector3d D);
         inline void setID(uint id) {m_id = id; };
 
         // TODO: check if acces by ref is good
-        inline std::vector<Eigen::Vector3f>& getCorners() {return m_corners; }; 
-        inline Eigen::Vector3f& getCornerA() {return m_corners[0]; };
-        inline Eigen::Vector3f& getCornerB() {return m_corners[1]; };
-        inline Eigen::Vector3f& getCornerC() {return m_corners[2]; };
-        inline Eigen::Vector3f& getCornerD() {return m_corners[3]; };
+        inline std::vector<Eigen::Vector3d>& getCorners() {return m_corners; }; 
+        inline Eigen::Vector3d& getCornerA() {return m_corners[0]; };
+        inline Eigen::Vector3d& getCornerB() {return m_corners[1]; };
+        inline Eigen::Vector3d& getCornerC() {return m_corners[2]; };
+        inline Eigen::Vector3d& getCornerD() {return m_corners[3]; };
         inline uint& getID() {return m_id; };
         open3d::geometry::TriangleMesh& getOpen3dMesh();
         Eigen::Vector3d& getUnorientedPlaneNormal();
@@ -61,17 +61,17 @@ namespace tslam
          * @param C corner C of the tag
          * @return Eigen::Vector3d The unoriented plane normal
          */
-        Eigen::Vector3d computeUnorientedPlaneNormal(const Eigen::Vector3f& A, const Eigen::Vector3f& B, const Eigen::Vector3f& C);
+        Eigen::Vector3d computeUnorientedPlaneNormal(const Eigen::Vector3d& A, const Eigen::Vector3d& B, const Eigen::Vector3d& C);
         /**
          * @brief Compute the center of the tag's polygon
          * 
          * @param A corner A of the tag
          * @return Eigen::Vector3d The center of the plane
          */
-        Eigen::Vector3d computeCenter(std::vector<Eigen::Vector3f> corners);
+        Eigen::Vector3d computeCenter(std::vector<Eigen::Vector3d> corners);
 
     private:
-        std::vector<Eigen::Vector3f> m_corners;
+        std::vector<Eigen::Vector3d> m_corners;
         uint m_id;
         open3d::geometry::TriangleMesh m_PlaneMesh;
         Eigen::Vector3d m_unorientedPlaneNormal;
